+++ src/components/User/Home/Products/ProductArea.tsx
@@ -5,7 +5,7 @@ import Link from "next/link";
 import Image from "next/image";
 import {TiEye} from 'react-icons/ti';
 import MenuProductArea from "@/components/User/Home/Products/MenuProductArea";
-import ProductQuickLookupModal from "@/components/User/ProductQuickLookupModal";
+import ProductQuickLookupModal from "@/components/User/ProductCommon/ProductQuickLookupModal";
 import useSWR from 'swr';
 import {getProductForUser, URL_API_PRODUCT_AREA} from "@/services/user/ProductAreaService";
