+++ src/css/globals.css
@@ -59,3 +59,5 @@ body {
     font-weight: bold;
 }
 
+
+
